1665 reformat code
