+++ monai/networks/nets/__init__.py
@@ -10,7 +10,6 @@
 # limitations under the License.
 
 from .ahnet import AHNet
-from .regunet import RegUNet
 from .autoencoder import AutoEncoder
 from .basic_unet import BasicUNet, BasicUnet, Basicunet
 from .classifier import Classifier, Critic, Discriminator
@@ -19,9 +18,8 @@ from .dynunet import DynUNet, DynUnet, Dynunet
 from .fullyconnectednet import FullyConnectedNet, VarFullyConnectedNet
 from .generator import Generator
 from .highresnet import HighResBlock, HighResNet
-from .localnet import LocalNet
 from .regressor import Regressor
-from .regunet import RegUNet
+from .regunet import LocalNet, RegUNet
 from .segresnet import SegResNet, SegResNetVAE
 from .senet import SENet, se_resnet50, se_resnet101, se_resnet152, se_resnext50_32x4d, se_resnext101_32x4d, senet154
 from .unet import UNet, Unet, unet
